Move VT parser into its own <vtparser> library and directory.
