+++ src/CMakeLists.txt
@@ -7,6 +7,7 @@ include(PedanticCompiler)
 add_subdirectory(crispy)
 add_subdirectory(text_shaper)
 add_subdirectory(vtpty)
+add_subdirectory(vtparser)
 add_subdirectory(terminal)
 add_subdirectory(terminal_renderer)
