+++ src/contour/display/TerminalWidget.cpp
@@ -30,8 +30,6 @@
 #include <fmt/chrono.h>
 #include <fmt/format.h>
 
-#include <range/v3/all.hpp>
-
 #include <QtCore/QDebug>
 #include <QtCore/QFileInfo>
 #include <QtCore/QFileSystemWatcher>
