+++ src/vtparser/test_main.cpp
@@ -1,0 +1,34 @@
+/**
+ * This file is part of the "libterminal" project
+ *   Copyright (c) 2019-2020 Christian Parpart <christian@parpart.family>
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include <crispy/App.h>
+#include <crispy/logstore.h>
+
+#define CATCH_CONFIG_RUNNER
+#include <catch2/catch.hpp>
+
+int main(int argc, char const* argv[])
+{
+    char const* logFilterString = getenv("LOG");
+    if (logFilterString)
+    {
+        logstore::configure(logFilterString);
+        crispy::App::customizeLogStoreOutput();
+    }
+    int const result = Catch::Session().run(argc, argv);
+
+    // avoid closing extern console to close on VScode/windows
+    // system("pause");
+
+    return result;
+}
